+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- fix window history state propagation in `popStateListener`
+
 ## [2.2.2] - 2023-10-13
 
 ### Fixed